Integrate AccountResetPage into AccountReset component
